BACKEND/requirements.txt
click==8.1.7
colorama==0.4.6
dlib==    19.24.6
face-recognition  ==      1.3.0
face_recognition_models== 0.3.0
mysql-connector-python == 9.0.0
numpy==   2.1.1
opencv-python        ==   4.10.0.84
pillow==  10.4.0
pip==     24.2